+++ tests/integration/test_integration.py
@@ -1,6 +1,7 @@
 import functools
 
 import pytest
+import sqlalchemy as sa
 
 import datajudge.requirements as requirements
 from datajudge.db_access import (
@@ -1942,11 +1943,13 @@ def test_max_null_fraction_between(engine, unique_table1, data):
         (identity, "col_varchar", "VARCHAR"),
         (identity, "col_int", "INTEGER"),
         (negation, "col_varchar", "INTEGER"),
+        (identity, "col_varchar", sa.types.String()),
+        (negation, "col_varchar", sa.types.Numeric()),
     ],
 )
 def test_column_type_within(engine, mix_table1, data):
     (operation, col_name, type_name) = data
-    if is_impala(engine):
+    if is_impala(engine) and type(type_name) == str:
         type_name = {"VARCHAR": "string", "INTEGER": "int"}[type_name]
     req = requirements.WithinRequirement.from_table(*mix_table1)
     req.add_column_type_constraint(col_name, type_name)