+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changelog
 **New features**
 
 - Implement :meth:`datajudge.WithinRequirement.add_value_distribution_constraint`.
+- Extended :meth:`datajudge.WithinRequirement.add_column_type_constraint` to support column type specification using string format, backend-specific SQLAlchemy types, and SQLAlchemy's generic types.
 
 
 1.6.0 - 2022.04.12